requirements.txt
asgiref==3.7.2
bcrypt==4.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.2.0
django-jazzmin==3.0.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
grpcio==1.66.1
grpcio-tools==1.66.1
gunicorn==21.2.0
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
pika==1.3.2
Pillow==10.0.0
protobuf==5.28.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
